add auto-survey-posting functionality
